+++ cluster-cm/src/main/java/com/sequenceiq/cloudbreak/cm/polling/ClouderaManagerCommandPollerObject.java
@@ -1,5 +1,7 @@
 package com.sequenceiq.cloudbreak.cm.polling;
 
+import java.math.BigDecimal;
+
 import com.cloudera.api.swagger.client.ApiClient;
 import com.sequenceiq.cloudbreak.domain.stack.Stack;
 
@@ -9,16 +11,16 @@ public class ClouderaManagerCommandPollerObject extends ClouderaManagerPollerObj
 
     private final ApiClient apiClient;
 
-    private final Integer id;
+    private final BigDecimal id;
 
-    public ClouderaManagerCommandPollerObject(Stack stack, ApiClient apiClient, Integer id) {
+    public ClouderaManagerCommandPollerObject(Stack stack, ApiClient apiClient, BigDecimal id) {
         super(stack, apiClient);
         this.stack = stack;
         this.apiClient = apiClient;
         this.id = id;
     }
 
-    public Integer getId() {
+    public BigDecimal getId() {
         return id;
     }
 }
